third_party/WebKit/LayoutTests/http/tests/dom/mutationobserver-parserappend-childlist-expected.txt

CONSOLE MESSAGE: line 8: PASS: childList mutation event fired from HTML parser change.
Test result is dumped on console.log. foo bar
